browser ui wrapper
