+++ shared/js/ui/base/index.es6.js
@@ -3,7 +3,7 @@ window.$ = window.jQuery = require('jquery')
 
 // base dependencies
 const mixins = require('./mixins/index.es6.js')
-const BaseModel = require('./$BROWSER-model.es6.js')
+const BaseModel = require('./model.es6.js')
 const BasePage = require('./page.es6.js')
 const BaseView = require('./view.es6.js')
