+++ shared/js/ui/pages/options.es6.js
@@ -6,7 +6,7 @@ const privacyOptionsTemplate = require('./../templates/privacy-options.es6.js')
 const WhitelistView = require('./../views/whitelist.es6.js')
 const WhitelistModel = require('./../models/whitelist.es6.js')
 const whitelistTemplate = require('./../templates/whitelist.es6.js')
-const BackgroundMessageModel = require('./../models/$BROWSER-background-message.es6.js')
+const BackgroundMessageModel = require('./../models/background-message.es6.js')
 const parseUserAgentString = require('./../models/mixins/parse-user-agent.es6.js')
 const renderFeedbackHref = require('./../templates/shared/render-feedback-href.es6.js')
 const renderBrokenSiteHref = require('./../templates/shared/render-broken-site-href.es6.js')
