+++ shared/js/ui/pages/popup.es6.js
@@ -7,7 +7,7 @@ const TopBlockedView = require('./../views/top-blocked-truncated.es6.js')
 const TopBlockedModel = require('./../models/top-blocked.es6.js')
 const topBlockedTemplate = require('./../templates/top-blocked-truncated.es6.js')
 const SiteView = require('./../views/site.es6.js')
-const SiteModel = require('./../models/$BROWSER-site.es6.js')
+const SiteModel = require('./../models/site.es6.js')
 const siteTemplate = require('./../templates/site.es6.js')
 const SearchView = require('./../views/search.es6.js')
 const SearchModel = require('./../models/search.es6.js')
@@ -15,7 +15,7 @@ const searchTemplate = require('./../templates/search.es6.js')
 const AutocompleteView = require('./../views/autocomplete.es6.js')
 const AutocompleteModel = require('./../models/autocomplete.es6.js')
 const autocompleteTemplate = require('./../templates/autocomplete.es6.js')
-const BackgroundMessageModel = require('./../models/$BROWSER-background-message.es6.js')
+const BackgroundMessageModel = require('./../models/background-message.es6.js')
 
 function Trackers (ops) {
   this.$parent = window.$('#popup-container')
