+++ shared/js/ui/views/tracker-networks.es6.js
@@ -1,7 +1,7 @@
 const ParentSlidingSubview = require('./sliding-subview.es6.js')
 const heroTemplate = require('./../templates/shared/hero.es6.js')
 const CompanyListModel = require('./../models/site-company-list.es6.js')
-const SiteModel = require('./../models/$BROWSER-site.es6.js')
+const SiteModel = require('./../models/site.es6.js')
 const trackerNetworksIconTemplate = require('./../templates/shared/tracker-network-icon.es6.js')
 const trackerNetworksTextTemplate = require('./../templates/shared/tracker-networks-text.es6.js')
